// src/app/models/shipmentFilterViewModel.ts
export class ShipmentFilterViewModel {
    OrderDateFrom: any;
    OrderDateTo: any;
    id: number;
    tableName: string;
    createdWhen: Date;
    createdBy: number;
    createdByFullName: string;
    createdByUsername: string;
    createdByCode: string;
    idDeleted: number;
    name: string;
    code: string;
    isLoop: boolean;
    PageSize: any;
    PageNumber: any;
    Cols: any;
}